Import zdoom's OPL emulators
zdoom offers three OPL emulators.
See http://zdoom.org/wiki/OPL_synth_emulation
Make use of their object-oriented interface.
This implies switching to float samples.

Pick the CPU-heavy but accurate JavaOPL for now, but this needs to be
configurable using a command-line switch eventually.

Source files were imported from zdoom git commit b285cbe.

+++ audioout.hh
@@ -8,7 +8,7 @@ void InitializeAudio(double AudioBufferLength, double OurHeadRoomLength);
 /** Start audio */
 void StartAudio();
 /** Send samples */
-void SendStereoAudio(unsigned long count, int* samples);
+void SendStereoAudio(unsigned long count, float* samples);
 /** Wait for samples to be consumed, apart from headroom length in seconds */
 void AudioWait();
 /** Shutdown audio system */
